+++ tests/SerializeTester.php
@@ -2,8 +2,11 @@
 <?php
 
 require_once __DIR__ . '/../vendor/autoload.php';
-require_once __DIR__ . '/../src/SerializeClass.php';
+//require_once __DIR__ . '/../src/Serialize.php';
 require_once __DIR__ . '/TestClass.php';
+use Serialize\SerializeJson;
+use Serialize\SerializeYaml;
+use Serialize\SerializeXml;
 
 $tester = new TestClass(
         'PHPA. Class serializer',
@@ -12,14 +15,38 @@ $tester = new TestClass(
         ['PHP', 'GIT', 'OOP', 'COMPOSER']
 );
 
-$JSON = new SerializeClass('JSON');
-$YAML = new SerializeClass('YAML');
-$XML = new SerializeClass('XML');
-$Script = new SerializeClass('Script');
+$JSON = new SerializeJson();
+$YAML = new SerializeYaml();
+$XML = new SerializeXml();
 
 
-echo 'JSON:' . \PHP_EOL . $JSON->serilalize($tester) . \PHP_EOL . \PHP_EOL;
-echo 'YAML:' . \PHP_EOL . $YAML->serilalize($tester) . \PHP_EOL . \PHP_EOL;
-echo 'XML:' . \PHP_EOL . $XML->serilalize($tester) . \PHP_EOL . \PHP_EOL;
-echo 'WRONG FORMAT: ' . \PHP_EOL . $Script->serilalize($tester) . \PHP_EOL . \PHP_EOL;
-echo 'WRONG DATA TYPE: ' . \PHP_EOL . $JSON->serilalize('SOME WRONG DATA STRING') . \PHP_EOL . \PHP_EOL;
+echo 'JSON ' . \PHP_EOL;
+try {
+    echo \PHP_EOL . $JSON->serilalize($tester) . \PHP_EOL . \PHP_EOL;
+} catch (\Exception $e){
+    echo $e->getMessage();
+}
+
+
+echo 'YAML: ' . \PHP_EOL;
+try {
+    echo \PHP_EOL . $YAML->serilalize($tester) . \PHP_EOL . \PHP_EOL;
+} catch (\Exception $e){
+    echo $e->getMessage();
+}
+
+
+echo 'XML: ' . \PHP_EOL;
+try {
+    echo \PHP_EOL . $XML->serilalize($tester) . \PHP_EOL . \PHP_EOL;
+} catch (\Exception $e){
+    echo $e->getMessage();
+}
+
+
+echo 'WRONG DATA TYPE: ' . \PHP_EOL;
+try {
+    echo \PHP_EOL . $JSON->serilalize('SOME WRONG DATA STRING');
+} catch (\Exception $e){
+    echo $e->getMessage();
+}